`leaflet-adapter` renamed from `map-utils`, moved to own directory
