+++ public/tutorial/accessible/script/decorative.js
@@ -2,7 +2,7 @@ import {
   addMarker,
   addTileLayer,
   createMap,
-} from '../../../script/map-utils.js';
+} from '../../../leaflet-adapter/map-utils.js';
 import {
   attributionOsm,
   urlTemplateOsm,
