+++ public/tutorial/accessible/script/interactive.js
@@ -1,8 +1,8 @@
 import {
   addMarker,
   addTileLayer,
-  createMap
-} from '../../../script/map-utils.js';
+  createMap,
+} from '../../../leaflet-adapter/map-utils.js';
 import {
   attributionOsm,
   urlTemplateOsm,
