+++ public/tutorial/choropleth/script/choropleth.js
@@ -4,7 +4,7 @@ import {
   addGeoJson,
   addTileLayer,
   createMap,
-} from '../../../script/map-utils.js';
+} from '../../../leaflet-adapter/map-utils.js';
 import {
   attributionOsm,
   urlTemplateOsm,
