+++ public/tutorial/custom-marker-icons/script/custom-marker-icons.js
@@ -2,7 +2,7 @@ import {
   addMarker,
   addTileLayer,
   createMap,
-} from '../../../script/map-utils.js';
+} from '../../../leaflet-adapter/map-utils.js';
 import {
   attributionOsm,
   urlTemplateOsm,
