+++ public/tutorial/quick-start/script/quick-start.js
@@ -5,7 +5,7 @@ import {
   addPopup,
   addTileLayer,
   createMap,
-} from '../../../script/map-utils.js';
+} from '../../../leaflet-adapter/map-utils.js';
 import {
   attributionOsm,
   urlTemplateOsm,
