+++ public/tutorial/wms/script/color-coded-hillshade.js
@@ -1,7 +1,7 @@
 import {
   createMap,
   wmsTileLayer,
-} from '../../../script/map-utils.js';
+} from '../../../leaflet-adapter/map-utils.js';
 
 wmsTileLayer({
   baseUrl: 'https://ows.mundialis.de/services/service?',
