+++ public/tutorial/zoom/script/delta.js
@@ -1,4 +1,4 @@
-import { createMap } from '../../../script/map-utils.js';
+import { createMap } from '../../../leaflet-adapter/map-utils.js';
 import { addCartoTileLayerWithZoomLevelControl } from './zoom.js';
 
 addCartoTileLayerWithZoomLevelControl(
