+++ public/tutorial/zoom/script/scale.js
@@ -3,7 +3,7 @@ import {
   addTileLayer,
   createMap,
   setMapView,
-} from '../../../script/map-utils.js';
+} from '../../../leaflet-adapter/map-utils.js';
 import {
   attributionCarto,
   urlTemplateCarto,
