+++ public/tutorial/zoom/script/zero.js
@@ -1,7 +1,7 @@
 import {
   addTileLayer,
   createMap,
-} from '../../../script/map-utils.js';
+} from '../../../leaflet-adapter/map-utils.js';
 import {
   attributionCarto,
   urlTemplateCarto,
